+++ portal-impl/src/com/liferay/portal/service/impl/OrgLaborLocalServiceImpl.java
@@ -28,7 +28,7 @@ public class OrgLaborLocalServiceImpl extends OrgLaborLocalServiceBaseImpl {
 
 	@Override
 	public OrgLabor addOrgLabor(
-			long organizationId, int typeId, int sunOpen, int sunClose,
+			long organizationId, long typeId, int sunOpen, int sunClose,
 			int monOpen, int monClose, int tueOpen, int tueClose, int wedOpen,
 			int wedClose, int thuOpen, int thuClose, int friOpen, int friClose,
 			int satOpen, int satClose)
@@ -69,7 +69,7 @@ public class OrgLaborLocalServiceImpl extends OrgLaborLocalServiceBaseImpl {
 
 	@Override
 	public OrgLabor updateOrgLabor(
-			long orgLaborId, int typeId, int sunOpen, int sunClose, int monOpen,
+			long orgLaborId, long typeId, int sunOpen, int sunClose, int monOpen,
 			int monClose, int tueOpen, int tueClose, int wedOpen, int wedClose,
 			int thuOpen, int thuClose, int friOpen, int friClose, int satOpen,
 			int satClose)
@@ -100,7 +100,7 @@ public class OrgLaborLocalServiceImpl extends OrgLaborLocalServiceBaseImpl {
 		return orgLabor;
 	}
 
-	protected void validate(int typeId) throws PortalException {
+	protected void validate(long typeId) throws PortalException {
 		listTypeService.validate(
 			typeId, ListTypeConstants.ORGANIZATION_SERVICE);
 	}
