+++ portal-impl/src/com/liferay/portal/service/impl/OrgLaborServiceImpl.java
@@ -29,7 +29,7 @@ public class OrgLaborServiceImpl extends OrgLaborServiceBaseImpl {
 
 	@Override
 	public OrgLabor addOrgLabor(
-			long organizationId, int typeId, int sunOpen, int sunClose,
+			long organizationId, long typeId, int sunOpen, int sunClose,
 			int monOpen, int monClose, int tueOpen, int tueClose, int wedOpen,
 			int wedClose, int thuOpen, int thuClose, int friOpen, int friClose,
 			int satOpen, int satClose)
@@ -78,7 +78,7 @@ public class OrgLaborServiceImpl extends OrgLaborServiceBaseImpl {
 
 	@Override
 	public OrgLabor updateOrgLabor(
-			long orgLaborId, int typeId, int sunOpen, int sunClose, int monOpen,
+			long orgLaborId, long typeId, int sunOpen, int sunClose, int monOpen,
 			int monClose, int tueOpen, int tueClose, int wedOpen, int wedClose,
 			int thuOpen, int thuClose, int friOpen, int friClose, int satOpen,
 			int satClose)
