+++ portal-impl/src/com/liferay/portal/service/impl/PhoneLocalServiceImpl.java
@@ -45,7 +45,7 @@ public class PhoneLocalServiceImpl extends PhoneLocalServiceBaseImpl {
 	@Override
 	public Phone addPhone(
 			long userId, String className, long classPK, String number,
-			String extension, int typeId, boolean primary)
+			String extension, long typeId, boolean primary)
 		throws PortalException {
 
 		return addPhone(
@@ -56,7 +56,7 @@ public class PhoneLocalServiceImpl extends PhoneLocalServiceBaseImpl {
 	@Override
 	public Phone addPhone(
 			long userId, String className, long classPK, String number,
-			String extension, int typeId, boolean primary,
+			String extension, long typeId, boolean primary,
 			ServiceContext serviceContext)
 		throws PortalException {
 
@@ -136,7 +136,7 @@ public class PhoneLocalServiceImpl extends PhoneLocalServiceBaseImpl {
 
 	@Override
 	public Phone updatePhone(
-			long phoneId, String number, String extension, int typeId,
+			long phoneId, String number, String extension, long typeId,
 			boolean primary)
 		throws PortalException {
 
@@ -178,7 +178,7 @@ public class PhoneLocalServiceImpl extends PhoneLocalServiceBaseImpl {
 
 	protected void validate(
 			long phoneId, long companyId, long classNameId, long classPK,
-			String number, String extension, int typeId, boolean primary)
+			String number, String extension, long typeId, boolean primary)
 		throws PortalException {
 
 		if (!PhoneNumberFormatUtil.validate(number)) {
