+++ portal-impl/src/com/liferay/portal/service/impl/WebsiteLocalServiceImpl.java
@@ -40,7 +40,7 @@ public class WebsiteLocalServiceImpl extends WebsiteLocalServiceBaseImpl {
 	@Deprecated
 	@Override
 	public Website addWebsite(
-			long userId, String className, long classPK, String url, int typeId,
+			long userId, String className, long classPK, String url, long typeId,
 			boolean primary)
 		throws PortalException {
 
@@ -51,7 +51,7 @@ public class WebsiteLocalServiceImpl extends WebsiteLocalServiceBaseImpl {
 
 	@Override
 	public Website addWebsite(
-			long userId, String className, long classPK, String url, int typeId,
+			long userId, String className, long classPK, String url, long typeId,
 			boolean primary, ServiceContext serviceContext)
 		throws PortalException {
 
@@ -130,7 +130,7 @@ public class WebsiteLocalServiceImpl extends WebsiteLocalServiceBaseImpl {
 
 	@Override
 	public Website updateWebsite(
-			long websiteId, String url, int typeId, boolean primary)
+			long websiteId, String url, long typeId, boolean primary)
 		throws PortalException {
 
 		validate(websiteId, 0, 0, 0, url, typeId, primary);
@@ -170,7 +170,7 @@ public class WebsiteLocalServiceImpl extends WebsiteLocalServiceBaseImpl {
 
 	protected void validate(
 			long websiteId, long companyId, long classNameId, long classPK,
-			String url, int typeId, boolean primary)
+			String url, long typeId, boolean primary)
 		throws PortalException {
 
 		if (!Validator.isUrl(url)) {
